\name{Five-package}
\alias{Five-package}
\alias{Five}
\docType{package}
\title{
\packageTitle{Five}
}
\description{
\packageDescription{Five}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{Five}

Maintainer: \packageMaintainer{Five}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = Five)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{Five}}
%\section{Documentation Index}{\packageIndices{Five}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
