% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapsAPI.R
\name{getdata_api}
\alias{getdata_api}
\title{Download and Parse JSON Data from a URL}
\usage{
getdata_api(url, filename)
}
\arguments{
\item{url}{A character string specifying the URL from which to download the data.}

\item{filename}{A character string specifying the name of the file to save the downloaded data.}
}
\value{
An R object containing the parsed JSON data.
}
\description{
This function downloads JSON data from a specified URL and saves it to a file.
It then reads the JSON content from the file and parses it into an R object.
}
\details{
The function uses the `downloader` package to handle the data download process and the `jsonlite` package to parse the JSON content. The downloaded JSON is read from the specified file and converted into an R object using `fromJSON`.
}
\examples{
\dontrun{
url <- "http://example.com/data.json"
filename <- tempfile(fileext = ".json")
data <- getdata_api(url, filename)
print(data)
}

}
